+++ R/read.cross.R
@@ -2,8 +2,8 @@
 #
 # read.cross.R
 #
-# copyright (c) 2000-2014, Karl W Broman
-# last modified June, 2014
+# copyright (c) 2000-2019, Karl W Broman
+# last modified Dec, 2019
 # first written Aug, 2000
 #
 #     This program is free software; you can redistribute it and/or
@@ -39,7 +39,7 @@ read.cross <-
              na.strings=c("-","NA"), genotypes=c("A","H","B","D","C"),
              alleles=c("A","B"), estimate.map=FALSE, convertXdata=TRUE,
              error.prob=0.0001, map.function=c("haldane", "kosambi", "c-f", "morgan"),
-             BC.gen = 0, F.gen = 0, crosstype, ...)
+             BC.gen = 0, F.gen = 0, crosstype=NULL, ...)
 {
     if(format == "csvrs") {
         format <- "csvsr"
@@ -147,11 +147,11 @@ read.cross <-
             class(cross$geno[[i]]) <- "X"
 
     # Fix up the X chromosome data for a backcross or intercross
-    chrtype <- sapply(cross$geno,class)
-    if(any(chrtype=="X") && convertXdata) {
-        if(class(cross)[1]=="bc")
+    chr_type <- sapply(cross$geno,chrtype)
+    if(any(chr_type=="X") && convertXdata) {
+        if(crosstype(cross)=="bc")
             cross <- fixXgeno.bc(cross)
-        if(class(cross)[1]=="f2") {
+        if(crosstype(cross)=="f2") {
             if(missing(alleles)) alleles <- c("A","B")
             cross <- fixXgeno.f2(cross, alleles)
         }
@@ -160,20 +160,12 @@ read.cross <-
     ## Pass through read.cross.bcsft for BCsFt (convert if appropriate).
     cross <- read.cross.bcsft(cross = cross, BC.gen = BC.gen, F.gen = F.gen, ...)
 
-    # re-estimate map?
-    if(estimate.map) {
-        cat(" --Estimating genetic map\n")
-        map.function <- match.arg(map.function)
-        newmap <- est.map(cross, error.prob=error.prob, map.function=map.function)
-        cross <- replace.map(cross, newmap)
-    }
-
     # store genotype data as integers
     for(i in 1:nchr(cross))
         storage.mode(cross$geno[[i]]$data) <- "integer"
 
     # check alleles
-    if(class(cross)[1] != "4way") {
+    if(crosstype(cross) != "4way") {
         if(length(alleles) > 2) {
             warning("length of arg alleles should be 2")
             alleles <- alleles[1:2]
@@ -199,17 +191,32 @@ read.cross <-
 
     attr(cross, "alleles") <- alleles
 
-    type <- class(cross)[1]
-    if(!missing(crosstype)) {
-        if(crosstype=="risib") cross <- convert2risib(cross)
-        else if(crosstype=="riself") cross <- convert2riself(cross)
-        else class(cross)[1] <- crosstype
+    if(is.null(crosstype)) crosstype <- crosstype(cross)
+    if(crosstype=="risib") cross <- convert2risib(cross)
+    else if(crosstype=="riself") cross <- convert2riself(cross)
+    else class(cross) <- c(crosstype, "cross")
+
+    # if 4-way cross, make the maps matrices
+    if(crosstype=="4way") {
+        for(i in seq_along(cross$geno)) {
+            if(!is.matrix(cross$geno[[i]]$map) || nrow(cross$geno[[i]]$map) < 2) {
+                cross$geno[[i]]$map <- rbind(cross$geno[[i]]$map, cross$geno[[i]]$map)
+            }
+        }
+    }
+
+    # re-estimate map?
+    if(estimate.map) {
+        cat(" --Estimating genetic map\n")
+        map.function <- match.arg(map.function)
+        newmap <- est.map(cross, error.prob=error.prob, map.function=map.function)
+        cross <- replace.map(cross, newmap)
     }
 
     # run checks
     summary(cross)
 
-    cat(" --Cross type:", class(cross)[1], "\n")
+    cat(" --Cross type:", crosstype(cross), "\n")
 
     cross
 }
@@ -226,8 +233,8 @@ fixXgeno.bc <-
     omitX <- FALSE
 
     # pull out X chr genotype data
-    chrtype <- sapply(cross$geno,class)
-    xchr <- which(chrtype=="X")
+    chr_type <- sapply(cross$geno,chrtype)
+    xchr <- which(chr_type=="X")
     Xgeno <- cross$geno[[xchr]]$data
 
     # find "sex" and "pgm" in the phenotype data
@@ -296,8 +303,8 @@ fixXgeno.f2 <-
     omitX <- FALSE
 
     # pull out X chr genotype data
-    chrtype <- sapply(cross$geno,class)
-    xchr <- which(chrtype=="X")
+    chr_type <- sapply(cross$geno,chrtype)
+    xchr <- which(chr_type=="X")
     Xgeno <- cross$geno[[xchr]]$data
 
     # find "sex" and "pgm" in the phenotype data