+++ R/read.cross.csv.R
@@ -2,8 +2,8 @@
 #
 # read.cross.csv.R
 #
-# copyright (c) 2000-2015, Karl W Broman
-# last modified Aug, 2015
+# copyright (c) 2000-2019, Karl W Broman
+# last modified Dec, 2019
 # first written Aug, 2000
 #
 #     This program is free software; you can redistribute it and/or
@@ -244,9 +244,8 @@ read.cross.csv <-
     class(cross) <- c(type,"cross")
 
     # check that nothing is strange in the genotype data
-    cross.type <- class(cross)[1]
-    if(cross.type=="f2") max.gen <- 5
-    else if(cross.type=="bc") max.gen <- 2
+    if(type=="f2") max.gen <- 5
+    else if(type=="bc") max.gen <- 2
     else max.gen <- 14
 
     # check that markers are in proper order
@@ -257,12 +256,6 @@ read.cross.csv <-
             cross$geno[[i]]$map <- cross$geno[[i]]$map[o]
             cross$geno[[i]]$data <- cross$geno[[i]]$data[,o,drop=FALSE]
         }
-    }
-
-    # if 4-way cross, make the maps matrices
-    if(type=="4way") {
-        for(i in 1:n.chr)
-            cross$geno[[i]]$map <- rbind(cross$geno[[i]]$map, cross$geno[[i]]$map)
     }
 
     # estimate genetic map